refactor(framework-dashboard): one notification engine for the two watched feeds
use-intervention-notifications and use-activity-notifications were the same
hook twice: the same WARMUP baseline with the same doc line, the same fire()
skeleton (permission gate, first-item guard, joined body, onclick + close),
and byte-identical effect bodies modulo which key/pick pair they called.
PR #935 unified the server half of this (dashboard/keys.ts); this is the
client half of the same move.

What genuinely differs per feed is wording and where a click goes, so that
is now a spec object; the two exported hooks keep their names and call
sites. The two test files fold together, which also collapses their copied
FakeNotification scaffold.

Also rewrites the stale comment on PROJECT_KEYS in lib/preferences.ts: it
still argued for the local copy #935 removed, while sitting on an import of
the shared list.
