+++ packages/framework-dashboard/pages/index/+Page.tsx
@@ -20,8 +20,7 @@ import { useRuns } from '../../lib/use-runs.js'
 import { useLoaded, usePolled } from '../../lib/use-async.js'
 import { useRoute } from '../../lib/use-route.js'
 import { useContextSet } from '../../lib/use-context-set.js'
-import { useInterventionNotifications } from '../../lib/use-intervention-notifications.js'
-import { useActivityNotifications } from '../../lib/use-activity-notifications.js'
+import { useActivityNotifications, useInterventionNotifications } from '../../lib/use-notifications.js'
 import { usePreferences, notificationsEnabled, newActivityEnabled, humanInterventionEnabled } from '../../lib/preferences.js'
 import { pendingChoices, agentViews } from '../../lib/live-state.js'
 import { useDocumentTitle } from '../../lib/document-title.js'
